test: refactor docker compose test
