+++ tests/docker_compose/conftest.py
@@ -1,8 +1,10 @@
 import os
+import subprocess
+import time
+from typing import List
 
 import docker
 import pytest
-
 from jina.logging.logger import JinaLogger
 
 client = docker.from_env()
@@ -65,3 +67,63 @@ def docker_images(request, image_name_tag_map):
         image_name + ':' + image_name_tag_map[image_name] for image_name in image_names
     ]
     return images
+
+
+class DockerComposeServices:
+    healthy_status = 'healthy'
+    unhealthy_status = 'unhealthy'
+
+    def __init__(self, dump_path, timeout_second=30):
+        self.dump_path = dump_path
+        self.timeout_second = timeout_second
+
+    def __enter__(self):
+        subprocess.run(
+            f'docker-compose -f {self.dump_path} up --build -d --remove-orphans'.split(
+                ' '
+            )
+        )
+
+        container_ids = (
+            subprocess.run(
+                f'docker-compose -f {self.dump_path} ps -q'.split(' '),
+                capture_output=True,
+            )
+                .stdout.decode("utf-8")
+                .split('\n')
+        )
+        container_ids.remove('')  # remove empty  return line
+
+        if not container_ids:
+            raise RuntimeError('docker-compose ps did not detect any launch container')
+
+        client = docker.from_env()
+
+        init_time = time.time()
+        healthy = False
+
+        while time.time() - init_time < self.timeout_second:
+            if self._are_all_container_healthy(container_ids, client):
+                healthy = True
+                break
+            time.sleep(0.1)
+
+        if not healthy:
+            raise RuntimeError('Docker containers are not healthy')
+
+    @staticmethod
+    def _are_all_container_healthy(
+            container_ids: List[str], client: docker.client.DockerClient
+    ) -> bool:
+
+        for id_ in container_ids:
+            status = client.containers.get(id_).attrs['State']['Health']['Status']
+
+            if status != DockerComposeServices.healthy_status:
+                return False
+        return True
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        subprocess.run(
+            f'docker-compose -f {self.dump_path} down --remove-orphans'.split(' ')
+        )
